requirements.txt
astroid==2.4.2
blinker==1.4
cachetools==4.2.1           
cffi==1.14.5
colorama==0.4.4            
Flask==1.1.2
Flask-Cors==3.0.10     
Flask-SQLAlchemy==2.4.4
Flask-Mail==0.9.1
joblib==1.0.1
lazy-object-proxy==1.4.3
mccabe==0.6.1
mysql-connector-python==8.0.22
oauthlib==3.1.0
paypalrestsdk==1.13.1
pika==1.2.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
pylint==2.6.0
pylint-flask==0.6
pylint-flask-sqlalchemy==0.2.0
pylint-plugin-utils==0.6
python-dateutil==2.8.1
pytz==2021.1
regex==2021.3.17
requests-oauthlib==1.3.0
SQLAlchemy==1.3.22
threadpoolctl==2.1.0
toml==0.10.1
tqdm==4.59.0
uritemplate==3.0.1
